+++ lib/node/pl-local/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @milaboratories/pl-local
 
+## 1.3.0
+
+### Minor Changes
+
+- 073c259: readability improvement & better robustness for mid-way pl downaload and unpack interruption
+
 ## 1.2.6
 
 ### Patch Changes