+++ lib/ui/uikit/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @milaboratories/uikit
 
+## 1.2.20
+
+### Patch Changes
+
+- db794fa: fix padding in the base dropdown
+
+## 1.2.19
+
+### Patch Changes
+
+- 60e38dc: Add PlDropdownRef component, update examples
+
 ## 1.2.18
 
 ### Patch Changes