+++ tests/config-local-ml-integration/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @milaboratories/config-local-ml-integration
 
+## 1.1.16
+
+### Patch Changes
+
+- Updated dependencies [073c259]
+  - @milaboratories/pl-local@1.3.0
+  - @milaboratories/pl-middle-layer@1.12.3
+
+## 1.1.15
+
+### Patch Changes
+
+- Updated dependencies [11d30ad]
+- Updated dependencies [11d30ad]
+  - @milaboratories/pl-middle-layer@1.12.2
+  - @milaboratories/pl-client@2.5.2
+
+## 1.1.14
+
+### Patch Changes
+
+- @milaboratories/pl-middle-layer@1.12.1
+
 ## 1.1.13
 
 ### Patch Changes